requirements.txt
fitz
nltk
sentence-transformers
faiss-cpu
fastapi
uvicorn
openai
numpy

